Add if statement to delete_comment view

+++ posts/views.py
@@ -60,5 +60,8 @@ def likes(request,slug):
 
 def delete_comment(request, comment_id):
     comment = get_object_or_404(Comment, id=comment_id)
-    comment.delete()
-    return redirect('posts')
+
+    if request.method == "POST":
+        comment.delete()
+        return redirect('posts')
+    return render(request, "comments/comments_delete.html", {'comment': comment})
